04_Tools/002_mlflow/02_mlflow_client/src/pyproject.toml
[project]
name = "02-mlflow-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlflow-skinny>=3.6.0",
]
